refactor: migrate all primitive imports to barrel re-exports
- Update 14 files that imported from primitive subfolders to use barrel
- Fix theme imports (useTheme, AccentColor, getColors) to use barrel
- Fix Icon type imports (IconName, IconVariant) to use barrel
- Fix SelectOption import to use barrel
- Fix ScrollContainer hardcoded border color to use useTheme()
- Update migration prep task
