+++ .agents/tasks/2026-03-15-primitives-migration-prep.md
@@ -86,15 +86,107 @@ When migrating to `quorum-shared`, primitives will be organized by platform:
 - [x] Restructured with two-layer palette/semantics architecture
 - [ ] **Note**: Colors will likely need further tuning once primitives are in `quorum-shared` and being tested on real mobile screens
 
-### 5. Documentation
+### 5. Documentation — DONE
 - [x] Updated all 6 primitives doc files to reflect audit changes
-- [ ] Update docs to remove Container references and reflect ModalContainer reorganization
+- [x] Updated docs to remove Container references and reflect ModalContainer reorganization
 
 ---
 
-## Context for Migration
+## Migration Execution Plan
 
-The primitives migration is part of the stacked PRs workflow:
+### Branches
+- **quorum-desktop**: `feat/shared-primitives-migration` (prep work committed)
+- **quorum-shared**: `feat/shared-primitives-migration` (created from `feat/shared-types-migration`)
+
+### Steps
+
+#### Step 1: Commit prep work on quorum-desktop — DONE
+- [x] All audit fixes, Container removal, type cleanup committed
+- [x] Branch: `feat/shared-primitives-migration`
+
+#### Step 2: Create branch on quorum-shared — DONE
+- [x] Branch `feat/shared-primitives-migration` created from `feat/shared-types-migration`
+- This ensures primitives migration builds on top of the types migration
+
+#### Step 3: Copy primitives to quorum-shared — DONE
+- [x] Copied 18 primitive folders + theme/ to `quorum-shared/src/primitives/`
+- [x] Removed SCSS imports (styles stay in consuming app)
+- [x] Fixed internal imports (logger → relative, self-references)
+- [x] Made business-layer deps injectable (UserAvatar → renderAvatar prop, processAvatarImage → onProcessImage prop)
+- [x] Added ReactTooltip wrapper inside Tooltip folder
+- [x] Fixed strict TypeScript errors for React 19 compatibility
+- [x] Removed @lingui/core — plain English defaults, apps pass translated strings
+- [x] Set up barrel exports (`src/index.ts` exports `./primitives`)
+- [x] Added peer dependencies (react-native, tabler icons, expo, clsx, etc.)
+- [x] 0 TypeScript errors in quorum-shared
+- [x] Committed: `d9f6678` (deps) + `ebe1b5a` (primitives)
+
+#### Step 4: Update quorum-desktop to use shared primitives
+**Approach**: Direct migration — no intermediate re-export layer.
+
+**4a. Rework `primitives/index.ts`**
+- [ ] Replace the barrel file with: SCSS imports + re-exports from `@quilibrium/quorum-shared`
+- [ ] All 12 SCSS files stay in their current folders (web bundler needs them)
+- [ ] The barrel imports them all, so they're loaded when any primitive is used
+- [ ] Re-exports only primitive components/types from `@quilibrium/quorum-shared`
+- This preserves all existing `import { X } from '../primitives'` paths — zero changes needed across the codebase
+
+**4b. Delete local primitive source files**
+- [ ] Remove all `.web.tsx`, `.native.tsx`, `types.ts`, `index.ts` from each primitive subfolder
+- [ ] Keep only `.scss` files in each subfolder
+- [ ] Keep `theme/` folder locally only if web app has custom ThemeProvider setup, otherwise delete
+- [ ] Keep `Icon/iconMapping.ts` and `Icon/icons/` if referenced directly by non-primitive code
+
+**4c. Fix Lingui translation defaults**
+Since quorum-shared primitives no longer use `t` macros, the web app must provide translated strings explicitly:
+- [ ] Audit all Select usages (~10) — add `placeholder={t\`Select an option\`}` where relying on default
+- [ ] Audit all FileUpload usages — add explicit error message props if needed
+- [ ] Audit `selectAllLabel` / `clearAllLabel` — add `t` wrapped props where multi-select is used
+
+**4d. Update mobile test screen imports**
+- [ ] Update imports in `mobile/test/primitives/` to reference `@quilibrium/quorum-shared`
+- [ ] These bypass the barrel file, so they need direct import updates
+
+**4e. Verify**
+- [ ] Run `npx tsc --noEmit` — zero new errors
+- [ ] Run `yarn build` — web app builds
+- [ ] Visual check: open web app, test Select dropdowns, FileUpload, modals
+- [ ] Commit
+
+#### Step 5: Local development linking
+- quorum-desktop already has `"@quilibrium/quorum-shared": "file:../quorum-shared"` in package.json
+- This means quorum-desktop will automatically use the local quorum-shared checkout
+- **Before merging**: switch back to published registry version
+
+### Commit Strategy
+- **quorum-shared**:
+  1. Separate commit for `package.json` changes (dependencies, peer deps, version bump)
+  2. One big commit for all primitives files (the actual migration)
+  3. Separate commits for any fixes, import adjustments, or non-primitive changes
+- **quorum-desktop**: Separate commits for import updates, cleanup, fixes
+- Keep commits minimal — the primitives migration itself is one atomic unit
+
+### Follow-up: Hardcoded Colors in Native Primitives
+Some native primitives still use hardcoded hex/rgba values instead of `useTheme()`. Fixed so far:
+- [x] Spacer — `'#e5e7eb'` → `theme.colors.border.default`
+- [x] ScrollContainer — `'rgba(255, 255, 255, 0.1)'` → `theme.colors.border.default`
+
+Still need review:
+- [ ] ColorSwatch — accent color map uses hardcoded hex values + check icon `color="#ffffff"`
+- [ ] Select.native — `color: '#fff'` (line 915) — white text, wrong in light mode
+- [ ] Tooltip.native — overlay `backgroundColor: 'rgba(0, 0, 0, 0.1)'`
+- Note: `shadowColor: '#000'` and modal backdrop `rgba(0, 0, 0, 0.5)` are OK to keep hardcoded (always dark)
+
+### 6. Remove Lingui from Shared Primitives — DONE
+Primitives in a shared library should not depend on a specific i18n framework.
+- [x] Replaced all `t` tagged template literals with plain English string defaults in FileUpload and Select
+- [x] Removed `@lingui/core/macro` imports from all 4 files (Select.web, Select.native, FileUpload.web, FileUpload.native)
+- [x] Removed `@lingui/core` from quorum-shared peerDependencies and peerDependenciesMeta
+- **Approach**: Consuming apps provide translated strings via props (e.g., `<Select placeholder={t\`Select an option\`} />`)
+- **Pattern**: Follows industry standard (shadcn/ui, Radix, MUI) — library uses English defaults, apps override with translations
+- **Action for consuming apps**: When using Select or FileUpload, pass `t\`...\`` wrapped props for any user-facing strings that need translation
+
+### Stacked PRs Workflow
 ```
 feat/shared-types-migration          ← PR #1
   └── feat/shared-primitives-migration      ← PR #2 (this work)
@@ -106,3 +198,4 @@ This prep task feeds into the actual migration plan at `tasks/2026-03-15-stacked
 ---
 
 _Created: 2026-03-15_
+_Updated: 2026-03-16_
